backend_docker/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.3
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.4.0
geocoder==1.38.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.0.4
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
netifaces==0.10.6
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.1
onnx-graphsurgeon==0.5.2
onnx2tf==1.20.0
onnxruntime==1.18.0
onnxsim==0.4.36
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openvino==2024.1.0
openvino-dev==2024.1.0
openvino-telemetry==2024.1.0
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pipenv==2023.12.1
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
py-machineid==0.5.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pydantic==1.10.15
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
sng4onnx==1.0.4
sounddevice==0.4.6
sparsezoo==1.7.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tflite-support==0.4.3
thop==0.1.1.post2209072238
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.5.0
tzdata==2024.1
ultralytics==8.2.20
urllib3==2.2.1
virtualenv==20.25.0
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.2
